Move the tilecreator to utils.
